position setup migration to global context // enabled in portfolio
