+++ src/components/Portfolio/index.tsx
@@ -20,371 +20,18 @@ import {
 import useFilteredPositions from "@hooks/useFilteredPositions";
 import useFilteredByChain from "@hooks/useFilteredByChain";
 import { useAtom } from "jotai";
-import { filteredChainAtom } from "@store/atoms";
+import { filteredChainAtom, positionsAtom } from "@store/atoms";
 import useFilteredPositionType from "@hooks/useFilteredPositionType";
 import { useQuery } from "@tanstack/react-query";
 import { fetchListicleFarms } from "@utils/api";
 import { FarmType } from "@utils/types";
 
-// Testing Object of position
-const positions = {
-  "moonbeam-stellaswap-0xF3a5454496E26ac57da879bf3285Fa85DEBF0388-38-axlDualPool":
-    {
-      unstaked: {
-        amount: 0,
-        amountUSD: 0,
-      },
-      staked: {
-        amount: 510.6994730234621,
-        amountUSD: 512.5126463600104,
-      },
-      unclaimedRewards: [
-        {
-          token: "STELLA",
-          amount: 20.164450219501695,
-          amountUSD: 20.164450219501695,
-        },
-        {
-          token: "WGLMR",
-          amount: 15.52313716327595,
-          amountUSD: 15.52313716327595,
-        },
-      ],
-    },
-  "moonbeam-curve-0xC106C836771B0B4f4a0612Bd68163Ca93be1D340-14-stDOT LP": {
-    unstaked: {
-      amount: 0.10615991769124165,
-      amountUSD: 0.5992796244723804,
-    },
-    staked: {
-      amount: 0,
-      amountUSD: 0,
-    },
-    unclaimedRewards: [
-      {
-        token: "LDO",
-        amount: 0,
-        amountUSD: 0,
-      },
-    ],
-  },
-  "moonbeam-curve-0x4efb9942e50aB8bBA4953F71d8Bebd7B2dcdE657-18-d2o-xcUSDT LP":
-    {
-      unstaked: {
-        amount: 0,
-        amountUSD: 0,
-      },
-      staked: {
-        amount: 2.105965635688815,
-        amountUSD: 2.106643574699173,
-      },
-      unclaimedRewards: [
-        {
-          token: "WGLMR",
-          amount: 0.04885905729800752,
-          amountUSD: 0.04885905729800752,
-        },
-      ],
-    },
-  "moonriver-solarbeam-0x0329867a8c457e9F75e25b0685011291CD30904F-6-WMOVR-xcKSM LP":
-    {
-      unstaked: {
-        amount: 3.3752569886e-7,
-        amountUSD: 0.010805298245182095,
-      },
-      staked: {
-        amount: 0,
-        amountUSD: 0,
-      },
-      unclaimedRewards: [
-        {
-          token: "SOLAR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-      ],
-    },
-  "moonriver-solarbeam-0x0329867a8c457e9F75e25b0685011291CD30904F-10-WMOVR-USDC LP":
-    {
-      unstaked: {
-        amount: 2.7230233464e-8,
-        amountUSD: 0.2146750120287031,
-      },
-      staked: {
-        amount: 0,
-        amountUSD: 0,
-      },
-      unclaimedRewards: [
-        {
-          token: "SOLAR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-      ],
-    },
-  "moonriver-solarbeam-0x0329867a8c457e9F75e25b0685011291CD30904F-9-FRAX-3pool":
-    {
-      unstaked: {
-        amount: 0.000996519600028022,
-        amountUSD: 0.00100718736901687,
-      },
-      staked: {
-        amount: 0,
-        amountUSD: 0,
-      },
-      unclaimedRewards: [
-        {
-          token: "SOLAR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-      ],
-    },
-  "moonriver-solarbeam-0x0329867a8c457e9F75e25b0685011291CD30904F-12-ETH-WMOVR LP":
-    {
-      unstaked: {
-        amount: 0.000001002834014023,
-        amountUSD: 0.00025896527140568,
-      },
-      staked: {
-        amount: 0,
-        amountUSD: 0,
-      },
-      unclaimedRewards: [
-        {
-          token: "SOLAR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-      ],
-    },
-  "moonriver-solarbeam-0x0329867a8c457e9F75e25b0685011291CD30904F-18-SOLAR-WMOVR LP":
-    {
-      unstaked: {
-        amount: 0.000542827971554763,
-        amountUSD: 0.00037704445086769495,
-      },
-      staked: {
-        amount: 0,
-        amountUSD: 0,
-      },
-      unclaimedRewards: [
-        {
-          token: "SOLAR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-      ],
-    },
-  "moonriver-solarbeam-0x0329867a8c457e9F75e25b0685011291CD30904F-25-kBTC-BTC":
-    {
-      unstaked: {
-        amount: 2.39272003623e-7,
-        amountUSD: 0.006322346311770977,
-      },
-      staked: {
-        amount: 0,
-        amountUSD: 0,
-      },
-      unclaimedRewards: [
-        {
-          token: "SOLAR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "xcKINT",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-        {
-          token: "WMOVR",
-          amount: 0,
-          amountUSD: 0,
-        },
-      ],
-    },
-  "moonriver-solarbeam-0x0329867a8c457e9F75e25b0685011291CD30904F-36-WMOVR": {
-    unstaked: {
-      amount: 2.3791017438e-8,
-      amountUSD: null,
-    },
-    staked: {
-      amount: 0,
-      amountUSD: null,
-    },
-    unclaimedRewards: [
-      {
-        token: "SOLAR",
-        amount: 0,
-        amountUSD: 0,
-      },
-    ],
-  },
-  "moonriver-solarbeam-0x0329867a8c457e9F75e25b0685011291CD30904F-20-BNB-BUSD LP":
-    {
-      unstaked: {
-        amount: 0.009961012323632836,
-        amountUSD: 0.3721775792041476,
-      },
-      staked: {
-        amount: 0,
-        amountUSD: 0,
-      },
-      unclaimedRewards: [
-        {
-          token: "SOLAR",
-          amount: 0,
-          amountUSD: 0,
-        },
-      ],
-    },
-  "moonriver-solarbeam-0x0329867a8c457e9F75e25b0685011291CD30904F-37-FRAX-WMOVR LP":
-    {
-      unstaked: {
-        amount: 0.013473274250934319,
-        amountUSD: 0.10845884385877262,
-      },
-      staked: {
-        amount: 0,
-        amountUSD: 0,
-      },
-      unclaimedRewards: [
-        {
-          token: "SOLAR",
-          amount: 0,
-          amountUSD: 0,
-        },
-      ],
-    },
-};
-
 const PortfolioPage = () => {
+  const [positions] = useAtom(positionsAtom);
+  useEffect(() => {
+    console.log("positions in portfolio", positions);
+  }, [positions]);
+
   // Avaiable chains we are supporting
   const chains = ["moonriver", "moonbeam", "astar", "mangata"];
   // Storage
@@ -415,7 +62,7 @@ const PortfolioPage = () => {
           address: farmInfo[2],
           id: farmInfo[3],
           lpSymbol,
-          ...value,
+          ...(value as any),
         };
       });
       // Filtering out the positions with null balance values
